svn-id: r43134

+++ sound/softsynth/mt32/partial.cpp
@@ -36,7 +36,7 @@
 // Hence we re-define them here. The only potential drawback is that it
 // might be a little bit slower this way.
 #define powf(x,y)	((float)pow(x,y))
-#define floorf(x)	((float)floorf(x))
+#define floorf(x)	((float)floor(x))
 #define fabsf(x)	((float)fabs(x))
 #endif
